#2107 editing dr location works now
