+++ src/backend/src/controllers/design-reviews.controllers.ts
@@ -78,8 +78,8 @@ export default class DesignReviewsController {
       const {
         dateScheduled,
         teamType,
-        requiredMembers,
-        optionalMembers,
+        requiredMembersIds,
+        optionalMembersIds,
         isOnline,
         isInPerson,
         zoomLink,
@@ -100,8 +100,8 @@ export default class DesignReviewsController {
         designReviewId,
         dateScheduled,
         teamType.teamTypeId,
-        requiredMembers,
-        optionalMembers,
+        requiredMembersIds,
+        optionalMembersIds,
         isOnline,
         isInPerson,
         zoomLink,
