+++ src/frontend/src/apis/design-reviews.api.ts
@@ -23,10 +23,10 @@ export const getAllDesignReviews = () => {
  * @param payload The new information for the design review
  */
 export const editDesignReview = (designReviewId: string, payload: EditDesignReviewPayload) => {
-  console.log('Editing design review' + designReviewId + payload.meetingTimes);
+  console.log(payload);
   return axios.post<{ message: string }>(apiUrls.designReviewsEdit(designReviewId), {
     ...payload
-  };
+  });
 };
 
 /**
